Extract id and time sorting parameters.
