+++ xchange-btce/src/main/java/com/xeiam/xchange/btce/v3/service/polling/trade/params/BTCETradeHistoryParams.java
@@ -14,15 +14,10 @@ import java.util.Date;
  *         Date: 4/2/15
  *         Time: 6:54 PM
  */
-public class BTCETradeHistoryParams extends DefaultTradeHistoryParamPaging implements TradeHistoryParamsIdSpan, TradeHistoryParamsTimeSpan,
+public class BTCETradeHistoryParams extends BTCETransHistoryParams implements
    TradeHistoryParamCurrencyPair {
 
   private CurrencyPair pair;
-  private BTCEAuthenticated.SortOrder sortOrder;
-  private String startId;
-  private String endId;
-  private Date startTime;
-  private Date endTime;
 
   @Override
   public void setCurrencyPair(CurrencyPair pair) {
@@ -36,53 +31,4 @@ public class BTCETradeHistoryParams extends DefaultTradeHistoryParamPaging imple
     return pair;
   }
 
-  @Override
-  public void setStartId(String startId) {
-
-    this.startId = startId;
-  }
-
-  @Override
-  public String getStartId() {
-
-    return startId;
-  }
-
-  @Override
-  public void setEndId(String endId) {
-
-    this.endId = endId;
-  }
-
-  @Override
-  public String getEndId() {
-    return endId;
-  }
-
-  @Override
-  public void setStartTime(Date startTime) {
-
-    this.startTime = startTime;
-  }
-
-  @Override
-  public Date getStartTime() {
-    return startTime;
-  }
-
-  @Override
-  public void setEndTime(Date endTime) {
-
-    this.endTime = endTime;
-  }
-
-  @Override
-  public Date getEndTime() {
-    return endTime;
-  }
-
-  public void setSortOrder(BTCEAuthenticated.SortOrder sortOrder) {
-
-    this.sortOrder = sortOrder;
-  }
 }
